+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next release
 
-- [CL-1058] - Add desktop preview
+- [CL-1058] - Add desktop preview in content builder
 
 ## 2022-07-04